+++ high_perf/common/data/gsmk_dataset.py
@@ -1,11 +1,8 @@
 from torch.utils.data import DataLoader
-import time
 from datasets import load_dataset
 
 # https://huggingface.co/datasets/gsm8k
 HUGGING_FACE_GSMK_DATASET_ID = "gsm8k"
-
-# Manual seed for reproducatibility
 
 def get_data_loader(batch_size, num_workers):
     dataset = load_dataset(HUGGING_FACE_GSMK_DATASET_ID, 'main')['train']
