+++ high_perf/common/model/get_model.py
@@ -1,0 +1,10 @@
+from llama import Llama
+
+def get_model(ckpt_dir, tokenizer_path, max_seq_len, max_batch_size):
+    generator = Llama.build(
+        ckpt_dir=ckpt_dir,
+        tokenizer_path=tokenizer_path,
+        max_seq_len=max_seq_len,
+        max_batch_size=max_batch_size,
+    )
+    return generator
